XMLHttpRequest/send-receive-utf16.htm

<!doctype html>
<meta charset=utf-8>
<title>XMLHttpRequest: The send() method: receive data which is UTF-16 encoded with BOM</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#document-response" data-tested-assertations="following::ol/li[9]" />
<div id="log"></div>

<script>
  async_test(function() {
    var client = new XMLHttpRequest();
    client.onload = this.step_func_done(function(e) {
      assert_equals(client.responseText, 'æøå\nテスト\n')
    });
    client.open("GET", "resources/utf16.txt");
    client.send(null);
  });
</script>
